fix: min_generation_interval_hours as float for 30m interval

+++ stock-auto-backend-python/app/strategy_evolution/models.py
@@ -16,7 +16,7 @@ class EvolutionConfig(BaseModel):
     fitness_return_weight: float = 0.5
     fitness_winrate_weight: float = 0.3
     fitness_mdd_penalty: float = 0.2
-    min_generation_interval_hours: int = 1
+    min_generation_interval_hours: float = 1.0
     elite_preserve_count: int = 5
     evolution_enabled: bool = True
     mdd_threshold: float = 10.0
